customised
made to use personally on linux
sudo python3 urlext.py
